requirements.txt
flask>=3.0
flask-cors>=4.0
supabase>=2.0
python-dotenv>=1.0
pydantic>=2.0
pytest>=7.4
pytest-flask>=1.3
openpyxl>=3.1
# LLM providers — install the one(s) you use
anthropic>=0.30.0
google-genai>=1.0.0
